==> sources/wiki/2_match_linked articles.ipynb <==
from lxml import etree
import pandas as pd
import swifter
import numpy as np
import re
import mwparserfromhell as mwp
import pandoc
from tqdm import notebook

LINK_REGEX = r"\[\[[\w\s\-]*[\|\w]+\]\]"

parse = etree.iterparse("data/wiki_coords_utf8.xml", huge_tree=True, tag='page')

dicts = []
for tag, page in parse:
    _dict = {}
    for node in page:
        _dict[node.tag] = node.text

    dicts.append(_dict)


df = pd.DataFrame(dicts)

df.loc[:, "title"] = df.title.str.lower()

df.loc[:, "lnkd_art"] = df.text.swifter.apply(lambda x: [match.lstrip("[").rstrip("]").lower().split("|") for match in re.findall(LINK_REGEX, x)])

df.loc[:, "lnkd_art_txt"] = df.lnkd_art.swifter.apply(lambda x: [(el[1].strip() if len(el)>1 else el[0].strip())for el in x])

df.loc[:, "lnkd_art_nms"] = df.lnkd_art.swifter.apply(lambda x: [el[0].strip() for el in x])

all_art_names = []
for art_names in df.lnkd_art_nms.values: all_art_names += art_names

unique_art_names = set(all_art_names)

unique_art_names = [name.strip() for name in unique_art_names]

art_intersection = set(unique_art_names).intersection(df.title.str.strip().to_list())

intersecter = set(art_intersection)

df.loc[:, "present_geo_arts"] = df.lnkd_art_nms.swifter.progress_bar(True).apply(lambda x: list(intersecter.intersection(x)))

df.loc[:, "parsed_text"] = df.text.swifter.apply(lambda x: mwp.parse(x).strip_code())

# df.to_csv('wiki_coords_article_article_matched.gz', chunksize=100000)
df.to_parquet("data/wiki_coords_article_matched.parquet")

df.head(20)
